Add del_rom register to stored calculator state
The del_rom register is used for goto and jsb instructions
to select the target rom bank. That register can be non-zero
when storing the state. After reloading the state on startup,
the del_rom will be initialized to 0. The next jump instruction
will then jump into the wrong bank, which can cause interesting
flickering display effects and a non-responsive simulator. The HP-45
was especially likely to produce this effect since the input loop
is in bank 3.